gettiing  close , moved component into folder, need to split css some styling then definitive styling
