+++ src/App.css
@@ -3,60 +3,3 @@
   padding: 0;
   box-sizing: border-box;
 }
-.card-container {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-}
-
-.card {
-  margin: 10px;
-  text-align: center;
-}
-
-.card img {
-  max-width: 100%;
-  height: 200px;
-  display: block;
-}
-.cardbox {
-  display: none;
-}
-.search-container {
-  display: flex;
-  justify-content: center;
-}
-
-@media only screen and (min-width: 768px) {
-  footer {
-    display: block;
-    background-color: var(--footer-background-color);
-    position: fixed;
-    bottom: 0;
-    width: 100%;
-  }
-
-  .footerContentContainer {
-    display: flex;
-    justify-content: space-evenly;
-    flex-wrap: nowrap;
-
-    h3 {
-      color: var(--text-color);
-      margin: 10px;
-    }
-  }
-  .cardbox {
-    display: block;
-  }
-  .searchresult-container {
-    display: flex;
-  }
-  .mainbody-container {
-    display: flex;
-  }
-  .search-container {
-    display: flex;
-    justify-content: flex-start;
-  }
-}
