--- src/components/DeckSection.jsx
@@ -1,15 +1,0 @@
-
-
-const DeckSection = () => {
-  return (
-        <div>
-          <div>deck</div>
-          <div>deck</div>
-          <div>deck</div>
-          <div>deck</div>
-        </div>
-  );
-};
-
-
-export default DeckSection;
