+++ src/components/decksection/DeckSection.jsx
@@ -1,0 +1,20 @@
+
+
+const DeckSection = () => {
+  return (
+        <div>
+          <div className="decksection">
+            <div>choose your deck</div>
+              <div className="decklist">
+                <div>deck elec</div>
+                <div>deck</div>
+                <div>deck</div>
+                <div>deck</div>
+              </div>
+            </div>
+          </div>
+  );
+};
+
+
+export default DeckSection;
