+++ src/components/mainbody/MainBody.css
@@ -1,0 +1,40 @@
+.card-container {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+  }
+  
+  .card {
+    margin: 10px;
+    text-align: center;
+  }
+  
+  .card img {
+    max-width: 100%;
+    height: 200px;
+    display: block;
+  }
+  .cardbox {
+    display: none;
+  }
+  .search-container {
+    display: flex;
+    justify-content: center;
+  }
+  
+  @media only screen and (min-width: 768px) {
+
+    .cardbox {
+      display: block;
+    }
+    .searchresult-container {
+      display: flex;
+    }
+    .mainbody-container {
+      display: flex;
+    }
+    .search-container {
+      display: flex;
+      justify-content: flex-start;
+    }
+  }
